Added a new operator (~=) to compare strings.
Now you should be able to do this:
"Louis-Ferdinand Céline" ~= "louis ferdinand celine" // true

By default it returns true if the strings are 40% different or less.
This value can be adjusted by changing the tolerance variable in the
operator definition.

I also added an example to the sample app, but I commented it out so as
to not interfere with the regular example.